requirements.txt
faiss-cpu
fastapi
grpcio
grpcio-tools
pandas
python-dotenv
requests
scikit-learn
torch
uvicorn